Fix PHP7 error with indexing a string as array
callHooks uses $extraInfo parameter as array, but some callers
passed '' here. This results in error in PHP7 when the function
tries to add $extraInfo['module'] value. For now I allowed this
parameter to be omitted and will initialize to empty array - needs
more testing and more modules may need update.