Remove error-causing ../MTN directory access.
This was used to read Monotone revision, but this should be
handled on a version control independent way, and on a way that
does not fail when open_base_dir restriction is in effect.
The code will now read a 'revision' file from site root, but there
are no tools yet to make that file. This whole _REV thing
could be removed instead and no nightly-build automatic upgrades.