Remove unused reference to self::getFileInfo()
This getFileInfo call used "module name", which is like "RSS" for
RSS theme, should use "osdirectory" there. Resulted in file not
found error and later on array index undefined (line 440).
But the resulting modBaseInfo array was not used anyway.
Also removed a commented out copy of same code later on.