Remove file tampering security checks of index.php
This function checks index.php and other main loader files
for a security hash. Removing now, so that they can be developed.
And need to decide if this check really makes sense or how to
update the the checksums.

Even simple acciendtal edit of index.php with vim breaks Xarigami
because you cannot remove the newline character from last line!