Fix error in config variable empty resultset check
ADOdb resultset is not null if there are no lines, this is a ResultSet
with 0 fields and 0 rows. BUT! Correcting this check resulted in a lot
of errors where caller didn't specify a default value, and this function
throws VariableNotFoundException!
A series of fixes will come to solve that.
The original code resulted in error in unserialze(null) on PHP 8!